+++ app/src/scenes/survey/daySurvey.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState, useContext, useRef, useMemo } from "react";
-import { ScrollView, View, KeyboardAvoidingView, Platform, Text } from "react-native";
-import { beforeToday, formatDate, formatDay, formatRelativeDate } from "../../utils/date/helpers";
-import { isToday, isYesterday, parseISO } from "date-fns";
+import { ScrollView, View, Text } from "react-native";
+import { beforeToday, formatDate, formatDay } from "../../utils/date/helpers";
 import { getScoreWithState } from "../../utils";
 import InputQuestion from "./InputQuestion";
 import logEvents from "../../services/logEvents";
@@ -32,10 +31,11 @@ import JMButton from "@/components/JMButton";
 import { useBottomSheet } from "@/context/BottomSheetContext";
 import HelpView from "@/components/HelpView";
 import { AnimatedHeaderScrollScreen } from "../survey-v2/AnimatedHeaderScrollScreen";
-import Pencil from "@assets/svg/Pencil";
 import NavigationButtons from "@/components/onboarding/NavigationButtons";
-import { interpolateColor, useAnimatedStyle, useSharedValue, withSpring } from "react-native-reanimated";
+import { useAnimatedStyle, useSharedValue, withSpring } from "react-native-reanimated";
 import ChevronIcon from "@assets/svg/icon/chevron";
+import { useStatusBar } from "@/context/StatusBarContext";
+import PencilIcon from "@assets/svg/icon/Pencil";
 
 const DaySurvey = ({
   navigation,
@@ -55,7 +55,8 @@ const DaySurvey = ({
     answers: {},
   };
   const { showBottomSheet, closeBottomSheet } = useBottomSheet();
-  const statusBarColorProgress = useSharedValue(0);
+  const { setCustomColor } = useStatusBar();
+  const animatedBackgroundColor = useSharedValue(TW_COLORS.PRIMARY);
   const [selectedMoodIndex, setSelectedMoodIndex] = useState<number | null>(null);
 
   const [diaryData, addNewEntryToDiaryData] = useContext(DiaryDataContext);
@@ -324,24 +325,16 @@ const DaySurvey = ({
     handleChangeUserComment({ key: getIndicatorKey(indicator), userComment: comment });
 
   const animatedStatusBarColor = useAnimatedStyle(() => {
-    if (selectedMoodIndex === null) {
-      return {
-        backgroundColor: TW_COLORS.PRIMARY,
-      };
-    }
-
-    const color = interpolateColor(statusBarColorProgress.value, [0, 0.25, 0.5, 0.75, 1], moodBackgroundColors);
-
     return {
-      backgroundColor: color,
+      backgroundColor: animatedBackgroundColor.value,
     };
   });
 
-  // Update statusBarColorProgress when selectedMoodIndex changes
+  // Update animatedBackgroundColor when selectedMoodIndex changes
   React.useEffect(() => {
     if (selectedMoodIndex !== null) {
-      const normalizedIndex = (selectedMoodIndex - 1) / (moodEmojis.length - 1);
-      statusBarColorProgress.value = withSpring(normalizedIndex);
+      animatedBackgroundColor.value = withSpring(moodBackgroundColors[selectedMoodIndex - 1]);
+      setCustomColor(moodBackgroundColors[selectedMoodIndex - 1]);
     }
   }, [selectedMoodIndex]);
 
@@ -354,7 +347,7 @@ const DaySurvey = ({
 
   return (
     <AnimatedHeaderScrollScreen
-      headerRightComponent={<Pencil color={selectedMoodIndex === null ? TW_COLORS.WHITE : TW_COLORS.CNAM_PRIMARY_900} width={16} height={16} />}
+      headerRightComponent={<PencilIcon color={selectedMoodIndex === null ? TW_COLORS.WHITE : TW_COLORS.CNAM_PRIMARY_900} width={16} height={16} />}
       headerRightAction={editIndicators}
       headerTitle={formatDate(initSurvey?.date)}
       handlePrevious={() => {