+++ app/src/utils/constants.ts
@@ -366,6 +366,7 @@ export const TW_COLORS = {
   GRAY_800: "#4A5D5F",
   GRAY_700: "#617778",
   GRAY_200: "#627577ff",
+  GRAY_500: "#92A9AB",
   GRAY_600: "#799092",
   GRAY_950: "#093F43",
   GRAY_50: "#F7FCFD",